# ARROW-9543: [C++] Simplify parsing/formatting utilities