… cpp/src/arrow/dataset/file_parquet_test.cc
… cpp/src/arrow/scalar_test.cc
… cpp/src/arrow/type_traits.h
… cpp/src/arrow/util/formatting.h
… cpp/src/arrow/util/formatting_util_test.cc
… cpp/src/arrow/util/uri.cc
… cpp/src/arrow/util/value_parsing.h
… cpp/src/arrow/util/value_parsing_benchmark.cc
… cpp/src/arrow/util/value_parsing_test.cc
… cpp/src/gandiva/gdv_function_stubs.cc