Replace `StringConverter` with `ParseValueTraits` and emphasize that the function `ParseValue` is the entrypoint for parsing primitive values.

Replace `StringFormatter` with `FormatValueTraits` and add a function `FormatValue` to act as the entrypoint for formatting primitive values.

A number of convenience overloads are provided for both.